requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9
slowapi>=0.1.9
anthropic>=0.34.0
google-genai>=1.0.0
tenacity>=8.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pymupdf>=1.24.0
markdown-it-py>=3.0.0
python-pptx>=1.0.0
google-cloud-storage>=2.17.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
redis>=5.0.0
python-dotenv>=1.0.0
